# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "OUTLIER API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/listings/{listing_id}": {
   "get": {
    "tags": [
     "listings"
    ],
    "summary": "Get Listing",
    "operationId": "get_listing_api_listings__listing_id__get",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "listings"
    ],
    "summary": "Patch Listing",
    "operationId": "patch_listing_api_listings__listing_id__patch",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ListingPatch"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "listings"
    ],
    "summary": "Delete Listing",
    "operationId": "delete_listing_api_listings__listing_id__delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/listings/{listing_id}/analyze": {
   "post": {
    "tags": [
     "analysis"
    ],
    "summary": "Analyze",
    "operationId": "analyze_api_listings__listing_id__analyze_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnalyzeIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/analysis": {
   "get": {
    "tags": [
     "analysis"
    ],
    "summary": "Analysis Runs",
    "operationId": "analysis_runs_api_listings__listing_id__analysis_get",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/identification": {
   "post": {
    "tags": [
     "analysis"
    ],
    "summary": "Correct Identification",
    "description": "User correction of the AI identification. Stored as a new current identification (origin=user).",
    "operationId": "correct_identification_api_listings__listing_id__identification_post",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IdentificationCorrection"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/identification/user": {
   "delete": {
    "tags": [
     "analysis"
    ],
    "summary": "Revert Identification",
    "operationId": "revert_identification_api_listings__listing_id__identification_user_delete",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/listings/{listing_id}/references": {
   "get": {
    "tags": [
     "analysis"
    ],
    "summary": "References For Listing",
    "operationId": "references_for_listing_api_listings__listing_id__references_get",
    "parameters": [
     {
      "name": "listing_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Listing Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AnalyzeIn": {
    "properties": {
     "mode": {
      "type": "string",
      "pattern": "^(auto|prefilter|triage|deep)$",
      "title": "Mode",
      "default": "auto"
     },
     "provider": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Provider"
     },
     "model": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Model"
     },
     "force": {
      "type": "boolean",
      "title": "Force",
      "default": false
     },
     "background": {
      "type": "boolean",
      "title": "Background",
      "default": false
     }
    },
    "type": "object",
    "title": "AnalyzeIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "IdentificationCorrection": {
    "properties": {
     "summary": {
      "type": "string",
      "maxLength": 500,
      "minLength": 2,
      "title": "Summary"
     },
     "domain": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Domain"
     },
     "confidence": {
      "anyOf": [
       {
        "type": "number",
        "maximum": 1.0,
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Confidence"
     },
     "brand_or_maker": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Brand Or Maker"
     },
     "era": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Era"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "demand_indicator": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Demand Indicator"
     },
     "liquidity_indicator": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Liquidity Indicator"
     }
    },
    "type": "object",
    "required": [
     "summary"
    ],
    "title": "IdentificationCorrection"
   },
...
   "ListingPatch": {
    "properties": {
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "category": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Category"
     },
     "seller": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Seller"
     },
     "domain": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Domain"
     },
     "condition_text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Condition Text"
     },
     "measurements": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Measurements"
     },
     "other_costs": {
      "anyOf": [
       {
        "additionalProperties": {
         "type": "number"
        },
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Other Costs"
     },
     "assumptions": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assumptions"
     },
     "user_notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Notes"
     },
     "archived": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archived"
     },
     "status": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Status"
     },
     "source_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Url"
     }
    },
    "type": "object",
    "title": "ListingPatch"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}